fix(process-instance): refine incident enrichment output #157
