+++ README.md
@@ -150,7 +150,7 @@ For batch execution:
 
 This is where `c8volt` becomes an operations tool instead of just a resource browser: it helps you see the process-instance structure that explains why a cancellation or deletion may behave the way it does.
 
-For incident diagnosis, add `--with-incidents` to keyed walks to show incident messages under the matching process-instance rows, or combine it with `--json` for structured incident details.
+For incident diagnosis, add `--with-incidents` to keyed walks to show incident keys and messages under the matching process-instance rows, or combine it with `--json` for structured incident details.
 
 ### Cancel Safely
 
@@ -244,7 +244,7 @@ For `get pd --stat`, Camunda `8.8` and `8.9` report process-instance counts for 
 
 Human process-instance lists mark only incident-bearing instances with `inc!`; instances without incidents omit the incident marker to keep long lists scannable.
 
-For direct keyed diagnosis, add `--with-incidents` to show incident error messages below the process-instance row, or combine it with `--json` for structured incident details. The flag is scoped to `--key` lookups; search filters such as `--incidents-only` keep their existing list-filter behavior.
+For direct keyed diagnosis, add `--with-incidents` to show incident keys and messages below the process-instance row, or combine it with `--json` for structured incident details. The flag is scoped to `--key` lookups; search filters such as `--incidents-only` keep their existing list-filter behavior.
 
 The `--start-date-*` and `--end-date-*` flags are inclusive `YYYY-MM-DD` bounds for search/list usage. Relative day filters use `--*-date-older-days N` for `N` days old or older and `--*-date-newer-days N` for `N` days old or newer.
