+++ specs/154-get-pi-incidents/tasks.md
@@ -39,16 +39,16 @@
 
 ## Phase 3: User Story 1 - Show Incident Messages for Keyed Lookup (Priority: P1) MVP
 
-**Goal**: `c8volt get pi --key <key> --with-incidents` shows incident error messages in human-readable output.
+**Goal**: `c8volt get pi --key <key> --with-incidents` shows incident keys and error messages in human-readable output.
 
-**Independent Test**: Run keyed command tests against fixture responses with one incident, multiple incidents, and no incidents, then verify human-readable output preserves the normal row and renders incident messages as indented `incident:` lines below the matching row.
+**Independent Test**: Run keyed command tests against fixture responses with one incident, multiple incidents, and no incidents, then verify human-readable output preserves the normal row and renders incident messages as indented `incident <incident-key>:` lines below the matching row.
 
 ### Tests for User Story 1
 
 - [x] T013 [P] [US1] Add `v88` service tests for process-instance incident search and error-message conversion in `internal/services/processinstance/v88/service_test.go`
 - [x] T014 [P] [US1] Add `v89` service tests for process-instance incident search and error-message conversion in `internal/services/processinstance/v89/service_test.go`
-- [x] T015 [P] [US1] Add command human-output test for one indented `incident:` message in `cmd/get_processinstance_test.go`
-- [x] T016 [P] [US1] Add command human-output tests for multiple indented `incident:` lines and no incidents in `cmd/get_processinstance_test.go`
+- [x] T015 [P] [US1] Add command human-output test for one indented `incident <incident-key>:` message in `cmd/get_processinstance_test.go`
+- [x] T016 [P] [US1] Add command human-output tests for multiple indented `incident <incident-key>:` lines and no incidents in `cmd/get_processinstance_test.go`
 
 ### Implementation for User Story 1
 
@@ -57,7 +57,7 @@
 - [x] T019 [US1] Add generated incident search method to `internal/services/processinstance/v89/contract.go`
 - [x] T020 [US1] Implement `v89` incident search request and conversion in `internal/services/processinstance/v89/service.go` and `internal/services/processinstance/v89/convert.go`
 - [x] T021 [US1] Call facade enrichment after keyed process-instance lookup when `--with-incidents` is set in `cmd/get_processinstance.go`
-- [x] T022 [US1] Add enriched human renderer for indented `incident:` message lines in `cmd/cmd_views_get.go`
+- [x] T022 [US1] Add enriched human renderer for indented `incident <incident-key>:` message lines in `cmd/cmd_views_processinstance_incidents.go`
 
 **Checkpoint**: User Story 1 is independently functional and testable.
 
@@ -77,8 +77,8 @@
 
 ### Implementation for User Story 2
 
-- [x] T026 [US2] Add enriched JSON renderer that emits `total` and per-item `incidents` in `cmd/cmd_views_get.go`
-- [x] T027 [US2] Ensure enriched JSON output preserves existing command envelope behavior in `cmd/cmd_views_get.go`
+- [x] T026 [US2] Add enriched JSON renderer that emits `total` and per-item `incidents` in `cmd/cmd_views_processinstance_incidents.go`
+- [x] T027 [US2] Ensure enriched JSON output preserves existing command envelope behavior in `cmd/cmd_views_processinstance_incidents.go`
 - [x] T028 [US2] Ensure empty incident results render as an empty collection when enrichment was requested in `c8volt/process/client.go`
 
 **Checkpoint**: User Stories 1 and 2 both work independently.
